// services/backend/src/workers/daily-content/schemas/types.ts
export const QuestionMetricTagSchema = z.object({
    question_id: z.string().uuid(),
    metric_keys: z.array(z.string()).max(2),
});

export const QuestionMetricTagArraySchema = z.array(QuestionMetricTagSchema);

export type QuestionMetricTag = z.infer<typeof QuestionMetricTagSchema>;

export type ReasoningGraphContext = {
    metric_keys: string[];
    nodes: {
        node_id: string;
        label: string;
        justification: string;
    }[];
    edges: {
        relationship: string;
        source_node_label: string;
        target_node_label: string;
    }[];
};

export type Node = {
    id: string;
    label: string;
    type: "ReasoningStep";
};